--- src/core/lib/transition/index.js
@@ -1,5 +1,0 @@
-export {default as TransitionInterpolator} from './transition-interpolator';
-export {default as ViewportFlyToInterpolator} from './viewport-fly-to-interpolator';
-export {default as LinearInterpolator} from './linear-interpolator';
-
-export {extractViewportFrom} from './transition-utils';
