+++ src/core/pure-js/README.md
@@ -1,0 +1,5 @@
+Pure JavaScript alternatives to React API.
+
+Still under development.
+
+Should be folded into lib folder.
