+++ src/core/transitions/transition-interpolator.js
