+++ src/core/utils/count.js
