+++ src/core/utils/flatten.js
