+++ src/deprecated-layers/extruded-choropleth-layer-64/extruded-choropleth-layer-64.js
@@ -19,8 +19,8 @@
 // THE SOFTWARE.
 
 import {Layer} from '../../core';
-import {fp64ify} from '../../core/lib/utils/fp64';
-import {flatten} from '../../core/lib/utils/flatten';
+import {fp64ify} from '../../core/utils/fp64';
+import {flatten} from '../../core/utils/flatten';
 import log from '../../core/utils/log';
 
 import {GL, Model, Geometry} from 'luma.gl';
